(V2) Concurrent Exchange<->Strategy synchronisation w/ messages
This PR introduces a WIP implementation of the `SyncAgent`, an Actor-based
component that is responsible for handling messages from both the strategy
and the exchange. Inbound data from the exchange - via the user stream - is
used to maintain a snapshot of the status of the account. Whilst messages
originating from the strategy are used to generate transactions destined
for the exchange. The SyncAgent runs within the context of it's own thread,
and communication is managed via a Queue and associated message types.

There are associated tests for the SyncAgent, written in Gherkin and using
the `behave` python package.

Incidental Work:

- All changes to the `AccountStatus` entity are now done via combined
update/mutation objects, available in the `mutations` module.
- Trader/Account and Strategy entities are now combined, this allows more
customisation when writing strategies, but also removes some moving parts.
- User defined `Strategy` objects are now expected to implement any risk
calculations - as opposed to having seperate `Calculator` objects.
- The Runner is now capable of handling graceful shutdowns, and terminating
the SyncAgent and API Adapter.
- Unit test coverage has improved, although the BDD tests need work for
compatibility with the new `SyncAgent`
- Various amendments and tweaks to interface/abstract classes.
- The `Makefile` now has a 'help' target.

As usual, these changes are driven by the living design document on notion.so